Buddify Prototype + Storyboard
This is a high-end prototype for the Buddify app. The app would, upon
implementation, match users up with other users, and set up a mutual gym
date. The main issue in society being: going to the gym is hard, but is
made easier if you have someone to go with. Users would be able to
choose who they wish to pair up. In other words, Tinder, but for
platonic gym friendships.

I was the project lead and was the principle designer.